Move utils into src/util
