+++ src/scripts/create.ts
@@ -3,10 +3,10 @@ import { join } from 'path';
 import { format } from 'prettier';
 import { Song } from '../definitions/song';
 import { Tag, TAGS } from '../definitions/tags';
-import generateFileName from '../generateFileName';
+import generateFileName from '../util/generateFileName';
 import { SONGS_FOLDER_PATH } from '../definitions/paths';
-import pushIds from '../pushIds';
-import { getAllSongPaths } from '../songs';
+import pushIds from '../util/pushIds';
+import { getAllSongPaths } from '../util/songs';
 
 export default function create(title: string, ...args: string[]): void {
 	if (!title) return console.error('A title for the new song must be provided!');
