+++ test/songFiles.test.ts
@@ -2,7 +2,7 @@ import { readFileSync } from 'fs';
 import * as matter from 'gray-matter';
 import { validMetaKeys } from '../src/definitions/song';
 import { TAGS } from '../src/definitions/tags';
-import { getAllSongPaths, getAllSongs } from '../src/songs';
+import { getAllSongPaths, getAllSongs } from '../src/util/songs';
 
 describe('All files in songs folder are songs', () => {
 	test('All files are markdown', () => {
